Move include tests into compiler
